secret-manager-example/database-secret-example/requirements.txt
psycopg2-binary==2.9.6
aioboto3==11.1.0